Clarify that multiple rows may contain lyrics
The text on a single folio may be present in multiple rows of the
exported CSV file. As the text alignment job discards syllables it
cannot match to text on the page, including the start of a chant that
continued onto the folio or the beginning of a chant that continues
after the folio is safe and handled by the job.

Co-authored-by: Timothy de Reuse <timdereuse@gmail.com>